qcacld-2.0: Implement descriptor pool for fw stats
The kernel address is used as cookie to keep track
of stats request. This address can be disclosed to
target leading to a security vulnerability.

Implement a FW stats descriptor pool, and use a
descriptor ID to keep track of stats requests,
instead of the kernel address, to prevent
kernel address leak.

Bug: 112277911
Test: Regression

CRs-Fixed: 2276007

+++ drivers/staging/qcacld-2.0/CORE/SERVICES/COMMON/ol_htt_api.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2011, 2014, 2016 The Linux Foundation. All rights reserved.
+ * Copyright (c) 2011, 2014, 2016, 2018 The Linux Foundation. All rights reserved.
  *
  * Previously licensed under the ISC license by Qualcomm Atheros, Inc.
  *
@@ -170,7 +170,7 @@ htt_h2t_dbg_stats_get(
     u_int32_t stats_type_reset_mask,
     u_int8_t cfg_stats_type,
     u_int32_t cfg_val,
-    u_int64_t cookie);
+    u_int8_t cookie);
 
 /**
  * @brief Get the fields from HTT T2H stats upload message's stats info header
